+++ changelog.rst
@@ -5,7 +5,8 @@ Change Log
 
 - In development.
 
-  - Fix bug inserting bookmarks into empty files.
+  - 2022/01/23: Fix error toggling a bookmark in a buffer without a file-name.
+  - 2022/01/19: Fix bug inserting bookmarks into empty files.
 
 - Version 0.1 (2022-01-02)